+++ CHANGELOG.md
@@ -111,7 +111,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 - Added new `EarlyStopping` parameters `stopping_threshold` and `divergence_threshold` ([#6868](https://github.com/PyTorchLightning/pytorch-lightning/pull/6868))
 
 
-- Added `BasePredictionWriter` Callback which enable to implement how to save predictions ([#7127](https://github.com/PyTorchLightning/pytorch-lightning/pull/7127))
+- Added `BasePredictionWriter` callback to implement prediction saving ([#7127](https://github.com/PyTorchLightning/pytorch-lightning/pull/7127))
 
 ### Changed